==> 3_SDM_CCVA.Rmd <==
---
title: "Use SDMs to measure CCV"
author: "Chloe Moore & TP DuBose"
date: "1/10/2023"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
#library(dismo)
library(raster); library(terra); library(stars)
library(landscapemetrics)
```

This is code to analyze SDM outputs in terms of habitat fragmentation, total expected area of occupancy, and potential realized climate niche breadth.

```{r bring in the data}
PATH_sdms<-'data/SDMs/'
PATH_spatext<-'spatial_extent.rds'
# need to add PRISM data if interested in quantifying climate niche breadth
prism.ppt<-''
prism.Tmax<-''
prism.Tmin<-''

# climate to project to
fut.clim<-''

# focal species
foc_species<-gsub('_topmodel.rds', '', list.files(PATH_sdms, pattern='topmodel.rds'))

crs.sdms<-"+proj=longlat +datum=WGS84 +no_defs" # crs of sdms
crs.albers <- st_crs("+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=km +no_defs")
```

Then we need to set up files that are consistently used to quantify climate change vulnerability among species.

```{r}
# prep the spatial extent to clip SDM rasters
se <- readRDS(PATH_spatext) %>% st_as_sf() %>%
  st_transform(crs.albers)
#st_crs(se)

# climate data for projection
fut<-rast(fut.clim)
fut<-crop(fut, ext(se %>% st_transform(st_crs(crs.sdms)))) #crop it to our spatial extent
names(fut)<-paste0('na_',gsub('o0','o', names(fut))) # make sure names match SDM models
fut 
#TODO: bring in elevation data and join to the future climate terra raster

# create raster of study extent
se_r<-st_rasterize(se, dx=100, dy=100) %>% #turn into a raster of 1x1 kilometers
  st_as_sf(as_points = FALSE, merge = FALSE) %>%  #turn into a polygon
  rowid_to_column() # add an id column
st_area(se_r[1,])
cat('raster created\n')
spat.ext.terra<-rast(se_r) # create a terra object instead of sf object

# cropping climate data
se_ppt<-terra::project(rast(prism.ppt), spat.ext.terra, mask=T)
se_tmx<-terra::project(rast(prism.Tmax), spat.ext.terra, mask=T)
se_tmn<-terra::project(rast(prism.Tmin), spat.ext.terra, mask=T)

stats_out<-NULL
for(spp in foc_species){
 # read in the outputs from the SDMs
  sp_files<-list.files(PATH_sdms, pattern=spp, full.names=T)
  print(sp_files[grepl('_bi_pred.asc', sp_files)])
  # read in the binary presense raster
  d.bin<-rast(sp_files[grepl('_bi_pred.asc',sp_files)])
  d.bin[d.bin==0] <- NA
  
  # project SDM to future climate 
  sdmout<-readRDS(sp_files[grepl('model.rds', sp_files)])
  sdmpred<-predict(sdmout, fut)
  # save projected climate out
  # threshold for binary output
  sdmpred[sdmpred < 0.6] <- NA # this is fake FIX ME!
  sdmpred[sdmpred >= 0.6] <- 1
  
 # for each raster type (binary current and future predicted)
  for(ras.type in c('d.bin', 'sdmpred')){
    sp.ras<-get(ras.type)
    clim_out<-NULL
    # get the realized niche breadth
    for(env.r in c('se_ppt','se_tmx','se_tmn')){
      env.ras<-get(env.r)
      env.ras<-project(env.ras, sp.ras, mask=T)
      env.var<-sp.ras*env.ras # assumes perfect alignment
      clim_out<-bind_rows(clim_out,
                          bind_cols(global(env.var, 'mean', na.rm=T),
                                    global(env.var, 'sd', na.rm=T)) %>%
                            mutate(clim=env.r,
                                   type=ras.type))}
    
    # landscape metrics on the binary raster
    spp_out<-bind_rows(
      lsm_l_ta(sp.ras), # total area in hectares
      lsm_l_cohesion(sp.ras), # patch cohesion
      lsm_l_area_mn(sp.ras), # mean area of patch
      # used in Hoecker & Turner et al. 2022
      lsm_l_enn_mn(sp.ras), # connectedness
      lsm_l_para_mn(sp.ras), # shape, perimieter:area
      # used in Rose et al. 2022
      lsm_p_area(sp.ras) %>% # number of patches
        group_by(layer, level, class, metric) %>%
        summarize(value=n(), .groups = 'drop'), 
      lsm_l_enn_cv(sp.ras)) %>% # patch isolation
      select(-layer, -id) %>%
      mutate(type=ras.type)
    
    # keep it all out and bind_rows to build summary data.frame
    stats_out <- bind_rows(stats_out, spp_out %>%
                             bind_rows(# realized climate niche breadth
                               clim_out %>% as_tibble() %>% 
                                 pivot_longer(-c(clim,type)) %>%
                                 mutate(metric=paste(clim, name, sep='.')) %>%
                                 select(-name, -clim)) %>%
                             mutate(species = gsub('_',' ',spp))) # add in species
  }
}
  
check_landscape(d.bin)
stats_out
write.csv(stats_out, 'data/analyzed_sdm_metrics.csv', row.names = F)
```

```{r}
list.files(PATH_sdms)

stats_out %>% pivot_wider(names_from=type) 
# calculate the RCS
rcs_index<-function(x)(1-(x-min(x))/(max(x)-min(x)))
rcs_sdms<-stats_out %>% filter(metric == 'ta' | grepl('.sd', metric)) %>%
  select(-level, -class) %>%
  pivot_wider(names_from=metric) %>% 
  group_by(type) %>%
  mutate(across(where(is.numeric), rcs_index), # scale
         CS=mean(c(se_ppt.sd, se_tmx.sd, se_tmn.sd)),
         RCS=(ta+CS)/2)
rcs_sdms %>% 
  ggplot()+geom_point(aes(x=RCS, y=species, color=type))
```
